moved: getMapChildElementByClassName
