+++ src/webgis4u/ol/util/getMapChildElementByClassName.test.js
