+++ src/webgis4u/util.js
@@ -5,5 +5,4 @@
 export * from './util/dom';
 export * from './util/function';
 export * from './util/number';
-export * from './util/ol';
 export * from './util/string';
